build: change submodule version logic
We now have to set the sha (commit id, tag) for the different
distros and branches explicitly in config/versions.

We need to update the submodule version manually, which is more
work, but it should be easier to focus on VDR*ELEC and not to fix
CoreELEC/LibreELEC bugs. Bugs can be bisected more easily, because
the CE/LE version is now tied to a specific VDR*ELEC commit.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>