rework cef build
Building the image with "-cef -extra cefbrowser" will package the large cef binaries
into the resulting image. It's located in /usr/local/config/cef-$ARCH.zip and has to be
unzipped within "/". The binaries will be in /storage/cefbrowser afterwards.
You have to do the same with /usr/local/config/web-cefbrowser.zip.
Due to the large filesize, the IMAGE_SIZE of the resulting image has to be increased
to 1024MB.

If the "-cef" is omitted and cefbrowser is built, the resulting zip isn't part of the
imgage. It is copied to the new subdirectory "cef" within the VDR*ELEC build dir
and has to be copied manually to the device.

Remember to clean the _cef package before you trigger builds with either "-cef" set
or not. LE/CE doesn't take notice of changed env variables and won't trigger a new
cef build.

TODO:
- Extend install.sh to make it easy to install cef and cefbrowser
- Combine that with remotetranscode
- Probably fix the systemd services and handle autostart